relasi produk & gambar

+++ routes/web.php
@@ -3,8 +3,10 @@
 use App\Http\Controllers\produkController;
 use App\Http\Controllers\UsersController;
 use App\Models\Category;
+use App\Models\kontak;
 use App\Models\Post;
 use App\Models\produk;
+use App\Models\profil;
 use App\Models\User;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\Route;
@@ -50,13 +52,16 @@ Route::get('/login', [UsersController::class, 'login'])->name('login')->middlewa
 Route::post('/login', [UsersController::class, 'auth_login'])->middleware('guest');
 Route::post('/logout', [UsersController::class, 'logout'])->name('logout')->middleware('auth');
 Route::get('/logout2', [UsersController::class, 'logout'])->name('logout2');
+//Add
 Route::get('/user/create', [UsersController::class, 'create'])->name('user.create')->middleware('auth');
-
-Route::post('/users', [UsersController::class, 'store'])->name('users.store');
-Route::get('/users', [UsersController::class, 'index'])->name('users');
-Route::delete('/users/{id}/destroy', [UsersController::class, 'destroy'])->name('users.destroy');
-Route::get('/users/{id}/edit', [UsersController::class, 'edit'])->name('users.edit');
+Route::post('/users', [UsersController::class, 'store'])->name('users.store')->middleware('auth');
+//delete
+Route::delete('/users/{id}/destroy', [UsersController::class, 'destroy'])->name('users.destroy')->middleware('auth');
+//edit update
+Route::get('/users/{id}/edit', [UsersController::class, 'edit'])->name('users.edit')->middleware('auth');
 Route::put('/users/update/{id}', [UsersController::class, 'update'])->name('users.update');
+//all
+Route::get('/users', [UsersController::class, 'index'])->name('users');
 
 Route::get('/user/detail/{produk:slug}', function (Produk $produk) {
     return view('user/detail', ['name' => 'Sandi Rp', 'title' => 'About', 'titles' => 'Halaman About', 'produk' => $produk]);
@@ -75,4 +80,7 @@ Route::get('/user/dashboard', function () {
     // $posts = Post::all();
     // return view('/user/dashboard', ['name' => 'Sandi Rp', 'title' => 'Blog', 'titles' => 'Halaman Blog', 'posts' => Post::filter(request(['search', 'category', 'author']))->latest()->paginate(10)->withQueryString()]);
     return view('/user/dashboard', ['name' => 'Sandi Rp', 'title' => 'Produks', 'titles' => 'Halaman Produk', 'produks' => produk::all()]);
+});
+Route::get('/user/kontak', function () {
+    return view('user/kontak', ['name' => 'Sandi Rp', 'title' => 'Kontak', 'titles' => 'Halaman Kontak', 'kontaks' => kontak::all(), 'profilP' => profil::all()]);
 });
